+++ app/Http/Controllers/MessageTypeController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Resources\Resource;
 use App\Models\MessageType;
+use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\Validator;
 use Symfony\Component\HttpFoundation\Response as ResponseCode;
 
@@ -13,6 +14,15 @@ class MessageTypeController extends Controller
     public function index(Request $request)
     {
         try {
+            $requestedUser = auth('api')->user();
+
+            if ($requestedUser == null) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'unautorized.',
+                ], ResponseCode::HTTP_UNAUTHORIZED);
+            }
+
             $messageTypes = MessageType::query();
 
             $messageTypes = $this->_fetchData($request, $messageTypes);
@@ -22,7 +32,7 @@ class MessageTypeController extends Controller
             _logger('error', 'MessageType', 'index', $e->getMessage());
 
             return response()->json([
-                'type' => 'error',
+                'status' => 'error',
                 'message' => 'server error 500.',
             ], ResponseCode::HTTP_INTERNAL_SERVER_ERROR);
         }
@@ -31,16 +41,34 @@ class MessageTypeController extends Controller
     public function show($messageType)
     {
         try {
-            $messageType = MessageType::withTrashed()
-            ->where('id', $messageType)
-            ->first();
+            $requestedUser = auth('api')->user();
+
+            if ($requestedUser == null) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'unautorized.',
+                ], ResponseCode::HTTP_UNAUTHORIZED);
+            }
+            $messageType = ($requestedUser->role == 'admin') ?
+                MessageType::withTrashed()
+                    ->where('id', $messageType)
+                    ->first() :
+                MessageType::where('id', $messageType)
+                    ->first();
+
+            if (!$messageType) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'message type not found.',
+                ], ResponseCode::HTTP_NOT_FOUND);
+            }
 
             return new Resource($messageType);
         } catch (\Exception $e) {
             _logger('error', 'MessageType', 'show', $e->getMessage());
 
             return response()->json([
-                'type' => 'error',
+                'status' => 'error',
                 'message' => 'server error 500.'
             ], ResponseCode::HTTP_INTERNAL_SERVER_ERROR);
         }
@@ -49,6 +77,15 @@ class MessageTypeController extends Controller
     public function store(Request $request)
     {
         try {
+            $requestedUser = auth('api')->user();
+
+            if ($requestedUser == null || $requestedUser->role != 'admin') {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'unautorized.',
+                ], ResponseCode::HTTP_UNAUTHORIZED);
+            }
+
             $validator = Validator::make($request->all(), [
                 'title' => 'required|string',
                 'description' => 'string',
@@ -57,7 +94,7 @@ class MessageTypeController extends Controller
 
             if ($validator->fails()) {
                 return response()->json([
-                    'type' => 'error',
+                    'status' => 'error',
                     'message' => $validator->errors(),
                 ], ResponseCode::HTTP_UNPROCESSABLE_ENTITY);
             }
@@ -66,14 +103,14 @@ class MessageTypeController extends Controller
             _logger('success', 'MessageType', 'store', $request->all());
 
             return response()->json([
-                'type' => 'success',
+                'status' => 'success',
                 'message' => 'message type created successfully.',
             ], ResponseCode::HTTP_CREATED);
         } catch (\Exception $e) {
             _logger('error', 'MessageType', 'store', $e->getMessage());
 
             return response()->json([
-                'type' => 'error',
+                'status' => 'error',
                 'message' => 'server error 500.',
             ], ResponseCode::HTTP_INTERNAL_SERVER_ERROR);
         }
@@ -82,6 +119,15 @@ class MessageTypeController extends Controller
     public function update(Request $request, $messageType)
     {
         try {
+            $requestedUser = auth('api')->user();
+
+            if ($requestedUser == null || $requestedUser->role != 'admin') {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'unautorized.',
+                ], ResponseCode::HTTP_UNAUTHORIZED);
+            }
+
             $validator = Validator::make($request->all(), [
                 'title' => 'string',
                 'description' => 'string',
@@ -90,7 +136,7 @@ class MessageTypeController extends Controller
 
             if ($validator->fails()) {
                 return response()->json([
-                    'type' => 'error',
+                    'status' => 'error',
                     'message' => $validator->errors(),
                 ], ResponseCode::HTTP_UNPROCESSABLE_ENTITY);
             }
@@ -103,14 +149,14 @@ class MessageTypeController extends Controller
             _logger('success', 'MessageType', 'update', $messageType);
 
             return response()->json([
-                'type' => 'success',
+                'status' => 'success',
                 'message' => "message type ($messageType->id) updated successfully.",
             ], ResponseCode::HTTP_OK);
         } catch (\Exception $e) {
             _logger('error', 'MessageType', 'update', $e->getMessage());
 
             return response()->json([
-                'type' => 'error',
+                'status' => 'error',
                 'message' => 'server error 500.',
             ], ResponseCode::HTTP_INTERNAL_SERVER_ERROR);
         }
@@ -119,18 +165,27 @@ class MessageTypeController extends Controller
     public function delete(MessageType $messageType)
     {
         try {
+            $requestedUser = auth('api')->user();
+
+            if ($requestedUser == null || $requestedUser->role != 'admin') {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'unautorized.',
+                ], ResponseCode::HTTP_UNAUTHORIZED);
+            }
+
             $messageType->delete();
             _logger('success', 'MessageType', 'delete', $messageType);
 
             return response()->json([
-                'type' => 'success',
+                'status' => 'success',
                 'message' => "message type ($messageType->title) and ($messageType->id) deleted successfully.",
             ], ResponseCode::HTTP_OK);
         } catch (\Exception $e) {
             _logger('error', 'MessageType', 'delete', $e->getMessage());
 
             return response()->json([
-                'type' => 'error',
+                'status' => 'error',
                 'message' => 'server error 500.',
             ], ResponseCode::HTTP_INTERNAL_SERVER_ERROR);
         }
@@ -139,18 +194,27 @@ class MessageTypeController extends Controller
     public function destroy(MessageType $messageType)
     {
         try {
+            $requestedUser = auth('api')->user();
+
+            if ($requestedUser == null || $requestedUser->role != 'admin') {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'unautorized.',
+                ], ResponseCode::HTTP_UNAUTHORIZED);
+            }
+
             $messageType->forceDelete();
             _logger('success', 'MessageType', 'destroy', $messageType);
 
             return response()->json([
-                'type' => 'success',
+                'status' => 'success',
                 'message' => "message type ($messageType->id) destroyed successfully.",
             ], ResponseCode::HTTP_OK);
         } catch (\Exception $e) {
             _logger('error', 'MessageType', 'destroy', $e->getMessage());
 
             return response()->json([
-                'type' => 'error',
+                'status' => 'error',
                 'message' => 'server error 500.',
             ], ResponseCode::HTTP_INTERNAL_SERVER_ERROR);
         }
@@ -159,18 +223,34 @@ class MessageTypeController extends Controller
     public function restore(MessageType $messageType)
     {
         try {
+            $requestedUser = auth('api')->user();
+
+            if ($requestedUser == null || $requestedUser->role != 'admin') {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'unautorized.',
+                ], ResponseCode::HTTP_UNAUTHORIZED);
+            }
+
+            if ($messageType->deleted_at == null) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'message type not found.',
+                ], ResponseCode::HTTP_NOT_FOUND);
+            }
+
             $messageType->restore();
             _logger('success', 'MessageType', 'restore', $messageType);
 
-            return response()->json([            
-                'type' => 'success',
+            return response()->json([
+                'status' => 'success',
                 'message' => "message type ($messageType->id) restored successfully.",
             ], ResponseCode::HTTP_OK);
         } catch (\Exception $e) {
             _logger('error', 'MessageType', 'restore', $e->getMessage());
 
             return response()->json([
-                'type' => 'error',
+                'status' => 'error',
                 'message' => 'server error 500.',
             ], ResponseCode::HTTP_INTERNAL_SERVER_ERROR);
         }
